apply: suspense to auth page
