+++ app/auth/page.tsx
@@ -1,20 +1,14 @@
-"use client";
-import React from "react";
-import { useSearchParams } from "next/navigation";
+import React, { Suspense } from "react";
+import CheckAuth from "@/components/auth/check-auth";
+const AuthFallback = () => {
+  return <>오 망함</>;
+};
 
 const AuthPage = () => {
-  const searchParams = useSearchParams();
-  const user_id = searchParams.get("user_id");
-  const access_token = searchParams.get("access_token");
-  const refresh_token = searchParams.get("refresh_token");
-  const is_new = searchParams.get("new");
   return (
-    <div>
-      <p>{user_id}</p>
-      <p>{access_token}</p>
-      <p>{refresh_token}</p>
-      <p>{is_new}</p>
-    </div>
+    <Suspense fallback={<AuthFallback />}>
+      <CheckAuth />
+    </Suspense>
   );
 };
